25 tests passing

+++ routes/notes.js
@@ -3,26 +3,30 @@
 const express = require('express');
 const router = express.Router();
 const Note = require('../models/note');
+const ObjectId = require('mongoose').Types.ObjectId;
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/', (req, res, next) => {
 
-  const searchTerm = req.query.searchTerm;
+  const {searchTerm, folderId, tagId }  = req.query;
   let filter = {};
-
   if (searchTerm) {
-    filter = { $or:[
-      {title:  { $regex: searchTerm }}, 
-      {content:{ $regex: searchTerm }},
-      {folderId: searchTerm }
-    ]};
+    filter.title = { $regex: searchTerm, $options: 'i' };
   }
+  if (folderId) {
+    filter.folderId = folderId;
+  }
+  if (tagId) {
+    filter.tagId = tagId;
+  }
+
   Note
   .find(filter)  
-  .sort({_id: 'asc'})
-  .then(results => {
+  .populate('tags')
+  .sort({ updatedAt: 'desc' })
+    .then(results => {
     if (results) {
       res.json(results);
-    } else{
+    } else {
       next(); //=>404 handler
     }
   })
@@ -31,8 +35,14 @@ router.get('/', (req, res, next) => {
 
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/:id', (req, res, next) => {
+  if (!ObjectId.isValid(req.params.id)) {
+    const err = new Error('Invalid id');
+    err.status = 400;
+    return next(err); // => Error handler
+  }
   Note
     .findById(req.params.id)
+    .populate('tags')
     .then(result => {
       if (result) {
         res.json(result); // => Client
@@ -46,7 +56,7 @@ router.get('/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 /* ========== POST/CREATE A NOTE ========== */
 router.post('/', (req, res, next) => {
-  const { title, content, folderId } = req.body;
+  const { title, content, folderId, tags } = req.body;
 
   /***** Never trust users - validate input *****/
   if (!title) {
@@ -54,11 +64,26 @@ router.post('/', (req, res, next) => {
     err.status = 400;
     return next(err); // => Error handler
   }
+  if (folderId && !ObjectId.isValid(folderId)) {
+    const err = new Error('The `folderId` is invalid!');
+    err.status = 400;
+    return next(err); // => Error handler
+  }
+  if (tags) {
+    tags.forEach((tag) => {
+      if (!ObjectId.isValid(tag)) {
+        const err = new Error('The `tagId` is invalid');
+        err.status = 400;
+        return next(err); // => Error handler
+      }
+    });
+  }
 
   const newNote = {
     title: title,
     content: content,
-    folderId: folderId
+    folderId: folderId,
+    tags: tags
   };
 
   Note
@@ -77,8 +102,7 @@ router.post('/', (req, res, next) => {
 
 /* ========== PUT/UPDATE A SINGLE NOTE ========== */
 router.put('/:id', (req, res, next) => {
-  const noteId = req.params.id;
-  const { title, content } = req.body;
+  const { id, title, content, folderId, tags } = req.body;
 
   /***** Never trust users - validate input *****/
   if (!title) {
@@ -86,14 +110,27 @@ router.put('/:id', (req, res, next) => {
     err.status = 400;
     return next(err); // => Error handler
   }
-
+  if (folderId && !ObjectId.isValid(folderId)) {
+    const err = new Error('`folderId` is not a valid Mongo ObjectId');
+    err.status = 400;
+    return next(err); // => Error handler
+  }
+  if (tags) {
+    tags.forEach((tag) => {
+      if (!ObjectId.isValid(tag)) {
+        const err = new Error('The `tagId` is invalid');
+        err.status = 400;
+        return next(err); // => Error handler
+      }
+    });
+  }
   const updateObj = {
     title: title,
     content: content
   };
 
   Note
-    .findByIdAndUpdate(noteId, {$set: updateObj}, { new: true })
+    .findByIdAndUpdate(id, {$set: updateObj}, { new: true })
     .then(result => {
       if (result) {
         res.json(result); // => Client
@@ -106,6 +143,11 @@ router.put('/:id', (req, res, next) => {
 
 /* ========== DELETE/REMOVE A SINGLE NOTE ========== */
 router.delete('/:id', (req, res, next) => {
+  if (!ObjectId.isValid(req.params.id)) {
+    const err = new Error('Invalid id');
+    err.status = 400;
+    return next(err); // => Error handler
+  }
   Note
     .findByIdAndRemove(req.params.id)
     .then(() => {
